pyproject.toml
[tool.poetry]
name = "mta-v600"
version = "0.1.0"
description = "Multi-tenant/Multi-app FastAPI orchestrator"
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
pydantic = "^2.0.0"
pydantic-settings = "^2.0.0"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"
httpx = "^0.28.1"
pyyaml = "^6.0"

# Local packages
cache-dsn = {path = "packages_py/cache_dsn", develop = true}
cache-request = {path = "packages_py/cache_request", develop = true}
cache-response = {path = "packages_py/cache_response", develop = true}
connection-pool = {path = "packages_py/connection_pool", develop = true}
fetch-client = {path = "packages_py/fetch_client", develop = true}
fetch-compose-cache-dsn = {path = "packages_py/fetch_compose_cache_dsn", develop = true}
fetch-compose-cache-request = {path = "packages_py/fetch_compose_cache_request", develop = true}
fetch-compose-cache-response = {path = "packages_py/fetch_compose_cache_response", develop = true}
fetch-compose-connection-pool = {path = "packages_py/fetch_compose_connection_pool", develop = true}
fetch-compose-rate-limiter = {path = "packages_py/fetch_compose_rate_limiter", develop = true}
fetch-compose-retry = {path = "packages_py/fetch_compose_retry", develop = true}
fetch-proxy-dispatcher = {path = "packages_py/fetch_proxy_dispatcher", develop = true}
fetch-rate-limiter = {path = "packages_py/fetch_rate_limiter", develop = true}
fetch-retry = {path = "packages_py/fetch_retry", develop = true}
static-config-property-management = {path = "packages_py/static_config_property_management", develop = true}
vault-file = {path = "packages_py/vault_file", develop = true}

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
httpx = "^0.28.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["packages_py", "fastapi_apps"]
python_files = "test_*.py"
python_functions = "test_*"
norecursedirs = ["__STAGE__", "__SPECS__", "__REVIEW__", ".git", "node_modules", ".venv", "dist", "build"]
addopts = "--ignore=__STAGE__ --ignore=__SPECS__ --import-mode=importlib"
asyncio_mode = "auto"
# Using importlib mode allows multiple test files with same basename (e.g., test_config.py)
# across different packages without namespace collisions
